+++ src/test/java/io/github/astrapi69/jaxb/ObjectToXmlConverterTest.java
@@ -1,8 +1,8 @@
 /**
  * The MIT License
- * <p>
+ *
  * Copyright (C) 2021 Asterios Raptis
- * <p>
+ *
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -10,10 +10,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- * <p>
+ *
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- * <p>
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -70,6 +70,8 @@ class ObjectToXmlConverterTest
 		xmlFile = new File(PathFinder.getSrcTestResourcesDir(), "club-jaxb.xml");
 		expected = RuntimeExceptionDecorator
 			.decorate(() -> ReadFileExtensions.readFromFile(xmlFile));
+		actual = actual.replace("\n", "").replace("\r", "");
+		expected = expected.replace("\n", "").replace("\r", "");
 		assertEquals(expected, actual);
 
 	}
@@ -92,6 +94,8 @@ class ObjectToXmlConverterTest
 		xmlFile = new File(PathFinder.getSrcTestResourcesDir(), "entry-jaxb.xml");
 		expected = RuntimeExceptionDecorator
 			.decorate(() -> ReadFileExtensions.readFromFile(xmlFile));
+		actual = actual.replace("\n", "").replace("\r", "");
+		expected = expected.replace("\n", "").replace("\r", "");
 		assertEquals(expected, actual);
 
 	}
@@ -130,6 +134,8 @@ class ObjectToXmlConverterTest
 		xmlFile = new File(PathFinder.getSrcTestResourcesDir(), "company-jaxb.xml");
 		expected = RuntimeExceptionDecorator
 			.decorate(() -> ReadFileExtensions.readFromFile(xmlFile));
+		actual = actual.replace("\n", "").replace("\r", "");
+		expected = expected.replace("\n", "").replace("\r", "");
 		assertEquals(expected, actual);
 	}