+++ src/test/java/io/github/astrapi69/jaxb/ObjectToXmlFileConverterTest.java
@@ -1,8 +1,8 @@
 /**
  * The MIT License
- * <p>
+ *
  * Copyright (C) 2021 Asterios Raptis
- * <p>
+ *
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -10,10 +10,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- * <p>
+ *
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- * <p>
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -26,6 +26,7 @@ package io.github.astrapi69.jaxb;
 
 import io.github.astrapi69.checksum.FileChecksumExtensions;
 import io.github.astrapi69.file.delete.DeleteFileExtensions;
+import io.github.astrapi69.file.read.ReadFileExtensions;
 import io.github.astrapi69.file.search.PathFinder;
 import io.github.astrapi69.throwable.RuntimeExceptionDecorator;
 import org.junit.jupiter.api.Assertions;
@@ -103,8 +104,8 @@ class ObjectToXmlFileConverterTest
 		converter.toXml(club, xmlFile);
 		actual = xmlFile;
 		expected = new File(PathFinder.getSrcTestResourcesDir(), "club-jaxb.xml");
-		assertEquals(FileChecksumExtensions.getChecksum(expected, true),
-			FileChecksumExtensions.getChecksum(actual, true));
+		assertEquals(ReadFileExtensions.fromFile(expected).replace("\n", "").replace("\r", ""),
+			ReadFileExtensions.fromFile(actual).replace("\n", "").replace("\r", ""));
 		RuntimeExceptionDecorator.decorate(() -> DeleteFileExtensions.delete(xmlFile));
 	}
 
@@ -126,8 +127,8 @@ class ObjectToXmlFileConverterTest
 		converter.toXml(company, xmlFile);
 		actual = xmlFile;
 		expected = new File(PathFinder.getSrcTestResourcesDir(), "company-jaxb.xml");
-		assertEquals(FileChecksumExtensions.getChecksum(expected, true),
-			FileChecksumExtensions.getChecksum(actual, true));
+		assertEquals(ReadFileExtensions.fromFile(expected).replace("\n", "").replace("\r", ""),
+			ReadFileExtensions.fromFile(actual).replace("\n", "").replace("\r", ""));
 		RuntimeExceptionDecorator.decorate(() -> DeleteFileExtensions.delete(xmlFile));
 	}
 }